Document the CommandLineController and rename Game to GameEngine
